requirements.txt
torch==2.0.1
torchvision==0.15.2
pandas==2.0.3
trimesh==3.22.3
numpy==1.23.4
einops==0.6.1
pyyaml==6.0
timm==0.9.2
imageio==2.31.1
scikit-learn==1.3.1
scipy==1.10.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tqdm==4.66.1